AS-FAQ-RAG/requirements.txt
fastapi
uvicorn
google-generativeai
openai
groq
sentence-transformers
faiss-cpu
requests
numpy
pandas
googletrans>=4.0.2
schedule
pyyaml